local_paper_analyzer/requirements.txt
arxiv>=2.2.0,<3
requests>=2.32.0
python-dotenv>=1.0.1
PyMuPDF>=1.24.0
tenacity>=8.4.1
PyYAML>=6.0.2
tqdm>=4.66.4
python-docx>=1.1.2
markdown>=3.6
